fixed build warning

+++ service/date-daemon.c
@@ -92,7 +92,7 @@ gboolean date_get_holiday (BusLunarDate *object,
 	GError *error = NULL;
 
 	daemon = DATE_DAEMON (user_data);
-	lunar_date_set_solar_date(daemon->date, year, month, day , hour, &error);
+	lunar_date_set_solar_date(daemon->date, year, month, day, (guint8) hour, &error);
 	if (error != NULL ) {
 		g_dbus_method_invocation_return_error (invocation, g_quark_from_static_string(DATE_DBUS_NAME), 1, "%s", error->message);
 		return FALSE;
